Phase 5.2 fixes: IRQ handling, networking, uptime, input
exceptions.S: re-enable IRQs during syscall handling. The EL0->EL1 SVC
exception masks DAIF.I automatically; without clearing it, the timer never
fires while the kernel is inside a syscall. This froze g_ticks, broke
task_sleep/cooperative scheduling, and left the keyboard and mouse
unresponsive. Added msr daifclr,#2 after EXCEPTION_ENTRY and msr daifset,#2
before EXCEPTION_EXIT. The re-mask before EXCEPTION_EXIT prevents a race
where a nested IRQ fires between the msr SPSR_EL1 write and eret, corrupting
the exception mode and causing an EL1 instruction abort on the user page.

virtio_net.c: add DSB barriers before QUEUE_ENABLE MMIO writes. ARM requires
DSB to order Normal-memory writes (descriptor/avail ring) before Device-memory
writes (MMIO). Without it QEMU can process QUEUE_ENABLE before seeing the RX
descriptors, so the RX queue is enabled with no buffers and DHCP OFFER never
arrives.

ip.c: fix IP destination filter during DHCP phase. When g_our_ip==0 the
filter must accept all packets so a unicast OFFER to the proposed address is
not silently dropped.

pci_ecam.c: detect transitional virtio-net (PCI 0x1000, subsys=1) in
addition to the modern form (0x1041). QEMU user-mode networking exposes the
transitional device.

dhcp.c: only overwrite subnet/gateway/DNS on ACK when the server provides
non-zero values, preserving fields already set from the OFFER.

arm_timer.c / arm_timer.h: add timer_seed_from_cntpct() which seeds g_ticks
from CNTPCT_EL0 just before enabling IRQs, so uptime counts from QEMU launch
rather than from the end of the boot sequence.

main.c: call timer_seed_from_cntpct() immediately before msr daifclr,#2.

printk.c: support zero-flag and width in format specifiers (%02x, %04lx).

virtio_blk.c: replace %llu with %lu; printk only implements %lu and on
AArch64 unsigned long is already 64-bit.

+++ kernel/drivers/block/virtio_blk.c
@@ -224,9 +224,9 @@ int virtio_blk_init(void)
     u32 cap_hi    = vcfg_r32(g_dc, 4u);
     g_num_sectors = ((u64)cap_hi << 32) | (u64)cap_lo;
 
-    kinfo("virtio-blk: %llu sectors (%llu MB)\n",
-          (unsigned long long)g_num_sectors,
-          (unsigned long long)(g_num_sectors / 2048));
+    kinfo("virtio-blk: %lu sectors (%lu MB)\n",
+          (unsigned long)g_num_sectors,
+          (unsigned long)(g_num_sectors / 2048));
 
     /* Set up request queue (index 0) */
     g_desc  = (virtq_desc_t  *)q_mem;
@@ -324,16 +324,16 @@ static int submit_request(u32 type, u64 lba, u32 num_sectors, u8 *data_buf)
     }
 
     if (g_used->idx == g_last_used) {
-        kwarn("virtio-blk: request timeout (lba=%llu type=%u)\n",
-              (unsigned long long)lba, (unsigned)type);
+        kwarn("virtio-blk: request timeout (lba=%lu type=%u)\n",
+              (unsigned long)lba, (unsigned)type);
         return -1;
     }
 
     g_last_used = g_used->idx;
 
     if (g_req_status != VIRTIO_BLK_S_OK) {
-        kwarn("virtio-blk: device error status=%u lba=%llu\n",
-              (unsigned)g_req_status, (unsigned long long)lba);
+        kwarn("virtio-blk: device error status=%u lba=%lu\n",
+              (unsigned)g_req_status, (unsigned long)lba);
         return -1;
     }
